development config for openshift, fixed a bug regarding project creation
Development configuration for ropneshift is added as
development-openshift.js file. When a mongodb service is created in
openshift, information including username, password, and database name is provided
as environment varibales. But they are in mongodb container. Also add
these environment variables to the container this app is running. The
mongodb connection string is like
mongodb://userV76:qceCOoBakBWAfyB8@172.30.178.132:27017/sampledb.
Moreover, abug is fixed. fsp.writeFileAsync function can not create a
file and then write to it if the parent directory is not yet created.
Now,the project directory is first created and then the needed project
files are added to it.