changing impact services
impact services are now as following: getNumberOfEndpoints,
listEndpoints, getEndpointDetails. The last one returns a corelationId
which the user needs to listen to it to get the data:
event.on(id, function(data){})
Since services are all promises, autogeneration of 'then' and 'catch' constructs are also
offered as separate services to developers.