App instanciation is done in the new way
Now in the UI, buttons for starting and stopping apps are removed.
In the new framework, there is no stoping and starting mechanism. App is
either in running state or in crashed state (when an error is thrown)
TO DO:
1- When app is crased, it is not shown immediately in the UI. The
current quick sollution is that when a user press 'see log' button,
apps' info in loaded (not a good solution).
2- Update app is not working with new mechanism.