build status is updated and deployment status is added while creating a new host. restart build functionality is alo added.
Creating a new host in OCP goes through several stages. It first creates a build,
if 'Complete', create a deployment, if 'Complete', creates the host (Pod).
The status of build and deployment stages is shown now in the host tab
while creating a host. It is possible that build fails. So, a button to
restart the build is added.