host redeployment functionality is added
while creating a host (RE) in OCP, it is possible that it fails. This
commit added a functionality to trigger the deployment again. In OCP 3.5
version the is an API to instantiate a deployment (AKA replication
contorller) from a deployment config
(POST /oapi/v1/namespaces/{namespace}/deploymentconfigs/{name}/instantiate).
How ever in OCP 3.3 version (which is used by french guys), this API is not impemented.
I used 'ChangeConfig' deployment trigger. So, I changed the deploynet
configuration a little bit (editting a useles environment variable) and
replace the old one by calling the API for replacing DC (PUT
/oapi/v1/namespaces/{namespace}/deploymentconfigs/{name})