container status is added to the table in the host tab
previously, host tab only showed the build and deployment status.
Sometimes, both build and deployment are successfull, but the container
(POD) creation encounters a problem. So I added the pod status to the
table.
NOTICE: Even sometimes the pod creation is successfull, but due to
unknown reasons (openshift network problems), hosts are not reachable.
In this case, hosts are shown red in the visualization tool. In this
sitations, User can go to the host tab and press redeply button.